branched UKF to make response version for BayesianAttractorModel
In the BAttM code I already have a UKFResponse.m which essentially does the same
thing. This now is for interoperability with the central functions in
BayesFilters and models. I branched it from UKF.m to be able to fetch later
updates, if necessary. 

I needed to make the criterion options more general. Hence, there is now a
criterion function which needs to be supplied. Therefore, this UKFResponse is
not compatible with the old UKFResponse in the original BAttM code.