stop inference when there is a nan in an observation
I did this to make model fitting faster: Inference now stops when the response
time of a subject is reached which I encoded by setting all subsequent dot
positions to nan. The assumption is that EPABC would reject samples anyway that
have a large difference between observed and simulated RT. Note also that the
nondecision time would also be added. The trial in which inference stops due to
nan are encoded as timeouts. I expect only moderate speed improvements, because
the procedure only cuts out trials with too long RTs relative to that of the
subject.

This modification does not change anything in the logic of the UKFs. There were
simply no nans in the observations previously. They would have also led to
errors.